Extrastereo plugin: increases linearly the difference between left and right channels (as the XMMS extrastereo plugin) which has some of "live" effect (use it to hear what I mean)
ex: mplayer media.avi -aop list=extrastereo[:mul=3.4]
The default coefficient (mul) is a float number that defaults to 2.5.  If you
set it to 0.0, you will have a mono sound (average of both channels), if you
set it to 1.0, you will have unchanged sound.

A patch for DOCS/sound.html is about to be sent to Gabucino.