Enable dvdkit on solaris
solaris has iconv() in libc.so, a separate libiconv.so library does not
exist.  Only add a -liconv library when we were succesful compiling a
test program with the -liconv option.

--enable-debug does not work on solaris, because test "-g" in configure
script fails with a test error message "test: argument expected". Use
string comparision to test for a non empty $_debug / $_profile