This patch contains bugfixes for the esd audio output driver that I uncovered while trying to send sound to a remote esd server over a wireless (11 mbs, just enough to handle to sound) link.
First, the sound was full "ticking" sounds.  I found a bug that
prevented the "send the remainder of this block" code from ever being
called - so large chunks of audio were simply being ignored.  Fixing
this bug removed the "ticking" from audio streams.

Fixing this bug, however, uncovered another problem - when the socket
buffer was full, doing a blocking write to finish the buffer would take
far too long and would turn video into a chunky mess.  I'd imagine this
blocking write would be fine for an audio-only stream, but it turns out
to hold up the video far too much.

The solution in this patch is to write as much data as possible to the
socket, and then return as soon as possible, reporting the number of
bytes actually written accurately back to mplayer.  I've tested it on
both local and remote esd servers, and it works well.

Patch by Benjamin Osheroff <ben@gimbo.net>