Remove unnecessary OSD visibility code
Don't set osd_visible directly in seek code. It's usually set to the
same value by set_osd_bar() already. In the cases where set_osd_bar()
is not called or doesn't set it (seek is triggered by EDL or loop, or
there is no video output) it seems to make no difference.

This fixes a minor bug: if you set OSD level to disabled while the
progress bar was visible then doing seeks before it disappeared would
keep it visible but not change its value.