libmpeg2: Remove local quantizer scale changes related to postprocessing.
Local changes to our locally forked library copy were incompatible with the
upstream version and prevented (easily) using the upstream version instead
of the forked copy. The md5sums of the decoded frames do not change for one
tested sample, so apparently the code is ineffectual and thus, after some
discussion with Luca, I decided that the changes are not worth keeping.