Make slabel work.
It is amazing that so far no skin is using a slabel (static label), but the
reason is quite simple. Former versions of the GUI crashed when rendering a
slabel.

The current GUI recognizes the malloc error but can't render the slabel
due to (probable) lack of memory.

An item that shall be fnt-rendered must have a valid item width.

Additionally, a break was missing which caused a slabel that would have been
already handled to erroneously run through the dlabel handling as well.