Make option '-noidle' working for the GUI.
So far, -idle was implicitly a default option to the GUI (although
the corresponding option variable was set to the '-noidle' value).

Explicitly set option '-idle' in the GUI now, remove the error (both
message and exit) that "the -idle option cannot be used with GMPlayer",
start the GUI without filename in either mode (idle or noidle) and
go to play_next_file only as long as the GUI has something to play
(or -idle was given).

This allows using option '-noidle' for the GUI which can be used to
quit it after it has finished playing.