vo direct3d: fix crash in osd drawing code
  The crash occurred when the video window was minimized and
the OSD should be drawn. This caused the OSD textures creation
to fail and later the OSD drawing code would assume they exist
and cause the crash.

  When the video window is minimized a size change event with
the new size 0x0 is reported. This is now handled by deferring
the textures recreation. If the OSD should be drawn in the
meantime it will be checked if the OSD textures have been
allocated.

  This commit is done with the permission of the vo direct3d
maintainer.

  This is based on patches from Roger Pack (rogerdpack2 (gmail))
and Patrick Dinh (patrickdinh33 (yahoo)).