Allow compilation with Libav
Some CPP Macros and codec ids are not (yet) available in libav, so use
them only if they are actually defined. This doesn't work for code ids,
as they are defined as enums. Therefore, #ifdefs tests for the presence
of the respective codec.

This approach should also allow to compile mplayer against earlier
versions of FFmpeg.