configure: Simplify Theora check.
The Theora check does some very elaborate tests to account for issues in
old libtheora versions.  Nowadays libtheora is stable and easy to test
for, so there is no longer a need for the extra complexity.

Also drop the Theora check that linked against tremor/bitwise.c.
Its purpose is now lost in the mists of time.