Add a file loaded by the GUI to its playlist.
The GUI operates on its playlist, so the file may not be missing there.

This fixes a blocking when trying to replay a file loaded by the GUI
(evLoad/evLoadPlay) instead of having been given as a command line
argument, as well as a segmentation fault when loading a file after a
playlist has been created.

Reported by Hans-Dieter Kosch, hdkosch kabelbw de.