Don't unnecessarily and annoyingly move the video window.
So far, this has been done at the start and end of a video (in
non-fullscreen mode) as well as when selecting half, normal or
double size of the video.

Don't position the video window to the values given in the skin file,
because the user might have moved it in the meantime, so just let it
where it currently is.

This closes Bugzilla #1355.