Fix issues with glibc on 32 bit.
Some glibc versions e.g. from RedHat will cause
incorrect behaviour when a certain magic symbol
is not encountered.
This is just abhorrent and simply unacceptable,
breaking well-behaving programs to the benefit
of ancient programs, but let's try to make it
work anyway (and hope they won't repeat the insanity).