freesdp/errorlist: Fix "bandwith" typo in string literal
This change was suggested because of the following lintian warnings:

 I: mencoder: spelling-error-in-binary usr/bin/mencoder bandwith bandwidth
 I: mplayer-gui: spelling-error-in-binary usr/bin/gmplayer bandwith bandwidth
 I: mplayer: spelling-error-in-binary usr/bin/mplayer bandwith bandwidth

The original patch fixed it also in other places. I did omit
those because these files are copied from another code base.

The changed error message is only used in function fsdp_strerror,
which is never used in MPlayer's code base.

Patch-By: Miguel A. Colón Vélez >debian.micove gmail com<