Add support for CD/(S)VCD/DVD image and DVD copy playback to the GUI.
Extend the file selector and the menu to load such files and utilize the
respective scheme (cd://, vcd://, dvd://) for the playback so that it
plays as if it were a real disc.

For CD/(S)VCD, libcdio is necessary in order to load the cue files.

For DVD, utilize the existing option to play files copied from a DVD or
DVD images.

Additionally, update the documentation.