configure: iconv: Remove hack that added -L/usr/lib
Seems not really a good idea to do such modifications to
the selection of the libraries that will be linked.

Also adding -L/usr/lib might invalidate previous results
of tests because now a different lib will finally be linked
in the end.

I got caught by this hack when compiling with a mingw compiler
and adding /usr/lib linker search path prioritized msys/cygwin
versions of the libs.