demux_ty: Try to fix calculation of used file size.
TiVo files can contain junk at the end, and this code
finds where the "good" data ends.
Unfortunately there was not a shred of sense in the original
code.
The code now seems to do at least something sensible with the
few files I had to test.

Fixes a warning from Coverity, which also thought this
code made no sense.